renamed tests that now use testcontainers
