+++ com.mycompany.app.shopCart/src/it/java/com/mycompany/shopcart/MVC/ModelViewControllerTestContainersIT.java
@@ -1,87 +1,81 @@
-package com.mycompany.shopcart.controller;
+package com.mycompany.shopcart.MVC;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.*;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 
-import java.util.List;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
-import org.assertj.swing.annotation.GUITest;
+import org.assertj.swing.core.matcher.JButtonMatcher;
+import org.assertj.swing.edt.GuiActionRunner;
+import org.assertj.swing.fixture.FrameFixture;
 import org.assertj.swing.junit.runner.GUITestRunner;
 import org.assertj.swing.junit.testcase.AssertJSwingJUnitTestCase;
 import org.bson.Document;
 import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
 import org.testcontainers.containers.MongoDBContainer;
 
 import com.mongodb.MongoClient;
 import com.mongodb.ServerAddress;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
+import com.mycompany.shopcart.controller.ShopController;
 import com.mycompany.shopcart.model.Product;
 import com.mycompany.shopcart.repository.mongo.ProductMongoRepository;
-import com.mycompany.shopcart.view.ShopView;
+import com.mycompany.shopcart.view.swing.ShopSwingView;
 
 @RunWith(GUITestRunner.class)
-public class ShopControllerRaceConditionIT extends AssertJSwingJUnitTestCase {
+public class ModelViewControllerTestContainersIT extends AssertJSwingJUnitTestCase {
 
 	@ClassRule
 	public static final MongoDBContainer mongo = new MongoDBContainer("mongo:4.4.3");
 
-	@Mock
-	private ShopView shopView;
-
 	private MongoClient mongoClient;
 	private ProductMongoRepository productRepository;
+	private FrameFixture window;
+	private ShopSwingView shopSwingView;
+	private ShopController shopController;
 	private MongoCollection<Document> productCollection;
-	private AutoCloseable closeable;
 
 	private static final String SHOP_DB_NAME = "shop";
 	private static final String PRODUCT_COLLECTION_NAME = "product";
 
 	@Override
 	protected void onSetUp() {
-		closeable = MockitoAnnotations.openMocks(this);
 		mongoClient = new MongoClient(
 				new ServerAddress(
 						mongo.getContainerIpAddress(), 
 						mongo.getFirstMappedPort()));
-		productRepository = 
-				new ProductMongoRepository(mongoClient, SHOP_DB_NAME, PRODUCT_COLLECTION_NAME);
+		productRepository = new ProductMongoRepository(mongoClient, SHOP_DB_NAME, PRODUCT_COLLECTION_NAME);
 		MongoDatabase database = mongoClient.getDatabase(SHOP_DB_NAME);
 		database.drop();
 		productCollection = database.getCollection(PRODUCT_COLLECTION_NAME);
+		window = new FrameFixture(robot(), GuiActionRunner.execute(() -> {
+			shopSwingView = new ShopSwingView();
+			shopController = new ShopController(shopSwingView, productRepository);
+			shopSwingView.setShopController(shopController);
+			return shopSwingView;
+		}));
+		window.show();
 	}
 
 	@Override
-	protected void onTearDown() throws Exception {
+	protected void onTearDown() {
 		mongoClient.close();
-		closeable.close();
 	}
 
 	@Test
-	@GUITest
-	public void testCheckoutWithConcurrentThreadsOnlyOneShouldCheckout() {
-		Product product = new Product("1", "test");
+	public void testCheckout() {
+		Product product = new Product("99", "availableProduct");
 		addTestProductToDatabase(product.getId(), product.getName());
-		List<Thread> threads = IntStream.range(0, 10)
-				.mapToObj(i -> new Thread(
-						() -> 
-						new ShopController(shopView, productRepository)
-							.checkoutProduct(product)))
-				.peek(t -> t.start())
-				.collect(Collectors.toList());
-		await().atMost(10, TimeUnit.SECONDS)
-		.until(() -> threads.stream().noneMatch(t -> t.isAlive()));
-		verify(shopView, times(1)).checkoutProduct(product);
-		verify(shopView, times(9))
-				.showErrorProductNotFound("The product you are trying to buy is no longer available", product);
+		GuiActionRunner.execute(() -> shopController.allProducts());
+		window.list("productList").selectItem(0);
+		window.button(JButtonMatcher.withText("Buy Selected")).click();
+		window.list("cartList").selectItem(0);
+		window.button(JButtonMatcher.withText("Checkout")).click();
+		await().atMost(5, TimeUnit.SECONDS)
+			.untilAsserted(() -> assertThat(productRepository.findById("99")).isNull());
 	}
 
 	private void addTestProductToDatabase(String id, String name) {
